Update templates to support marking target word with `<strong>`
Anki 2.1.50 brought a change to the WYSIWYG note editor.  Previously,
pressing `Control+b` or the `Bold text` button wrapped the selection in
a `<b>` tag, and the latest update changed it to `<strong>`, therefore
breaking scripts and styles for newly added notes.  This patch adds
support for the new tag in affected templates.